- make authentication class has 2 behaviors only login and logout - move dependency of cacheRepository from logic layer to data layer - create MateDataSource for add or delete or edit mate - delete SignupUseCase.kt as it is not used - delete ClearLoggedInUserFromCacheUseCase.kt as it behave as LogoutUseCase - refactor test cases - add test cases for CredentialValidator
