+++ src/main/kotlin/di/logicModule.kt
@@ -7,10 +7,10 @@ import org.koin.dsl.module
 
 val logicModule = module {
     single { GetLogsByEntityIdUseCase(get()) }
-    single { ClearLoggedInUserFromCacheUseCase(get()) }
     single { GetLoggedInUserUseCase(get()) }
     single { ProjectUseCases(get(), get(), get()) }
-    single { LoginUseCase(get(), get(), get()) }
+    single { LogoutUseCase(get()) }
+    single { LoginUseCase(get()) }
     single { ManageTaskUseCase(get(), get(), get()) }
     single { ManageStateUseCase(get(), get(), get()) }
     single { CreateMateUseCase(get(), get(), get()) }
