--- src/main/kotlin/logic/useCases/ClearLoggedInUserFromCacheUseCase.kt
@@ -1,11 +1,0 @@
-package logic.useCases
-
-import logic.repositories.CacheDataRepository
-
-class ClearLoggedInUserFromCacheUseCase(
-    private val cacheDataRepository: CacheDataRepository
-) {
-    fun clearLoggedInUserFromCache() {
-        cacheDataRepository.clearLoggedInUserFromCache()
-    }
-}
