+++ src/main/kotlin/logic/useCases/CreateMateUseCase.kt
@@ -2,13 +2,13 @@ package logic.useCases
 
 import logic.entities.User
 import logic.exceptions.*
-import logic.repositories.AuthenticationRepository
 import logic.repositories.CacheDataRepository
+import logic.repositories.MateRepository
 import logic.validation.CredentialValidator
 
 class CreateMateUseCase(
     private val cacheDataRepository: CacheDataRepository,
-    private val authenticationRepository: AuthenticationRepository,
+    private val mateRepository: MateRepository,
     private val credentialValidator: CredentialValidator
 ) {
     fun createMate(username: String, password: String) {
@@ -25,13 +25,13 @@ class CreateMateUseCase(
         credentialValidator.takeIfValidNameOrThrowException(username)
         credentialValidator.takeIfValidPasswordOrThrowException(password)
 
-        authenticationRepository.getMates()
+        mateRepository.getMates()
             .any { it.userName == username }
             .takeIf { it }?.let {
                 throw UsernameTakenException()
             }
         val registered = try {
-            authenticationRepository.register(username, password)
+            mateRepository.addMate(username, password)
         } catch (e: UserAlreadyExistException) {
             throw UsernameTakenException()
         } catch (e: Exception) {
