+++ src/main/kotlin/logic/useCases/LogoutUseCase.kt
@@ -1,15 +1,11 @@
 package logic.useCases
 
 import logic.repositories.AuthenticationRepository
-import logic.repositories.CacheDataRepository
 
 class LogoutUseCase(
-    private val authenticationRepository: AuthenticationRepository,
-    private val cacheDataRepository: CacheDataRepository
+    private val authenticationRepository: AuthenticationRepository
 ) {
     operator fun invoke(): Boolean {
-        if (!authenticationRepository.logout()) return false
-        cacheDataRepository.clearLoggedInUserFromCache()
-        return true
+        return authenticationRepository.logout()
     }
 }
